Updated the speed lines to directly update based on the velocity of the plane.
The speed lines length and spawn rate are linked to the calculated air speed of the aircraft in order to modify the speed lines length and rate of spawning.

Further modification needed to fine tune parameters, along with adding the change to the wind lines speed of movement over the camera plane.